requirements.txt
absl-py==2.3.1
acvl_utils==0.2.5
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
asyncssh==2.22.0
atpublic==7.0.0
attrs==25.4.0
babel==2.17.0
batchgenerators==0.25.1
batchgeneratorsv2==0.3.0
beautifulsoup4==4.14.3
billiard==4.2.4
bleach==6.3.0
blosc2==3.12.2
bs4==0.0.2
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.3
configobj==5.0.9
connected-components-3d==3.26.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.25.2
dvc==3.66.1
dvc-data==3.18.2
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
dynamic_network_architectures==0.4.2
einops==0.8.1
entrypoints==0.4
executing==2.2.1
fastjsonschema==2.21.2
fft-conv-pytorch==1.2.0
filelock==3.20.0
flatten-dict==0.4.2
flufl.lock==9.0.0
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.12.0
funcy==2.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.46
grandalf==0.8
graphviz==0.21
gto==1.9.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.1
hydra-core==1.3.2
idna==3.11
imagecodecs==2026.1.1
ImageIO==2.37.2
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
iterative-telemetry==0.0.10
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
-e git+https://github.com/neheller/kits23.git@5f2b8cd369df7936d3253201560ba831b343ce9f#egg=kits23
kiwisolver==1.4.9
kombu==5.6.2
lark==1.3.1
lazy_loader==0.4
linecache2==1.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
narwhals==2.15.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
ndindex==1.10.1
nest-asyncio==1.6.0
networkx==3.6.1
nibabel==5.3.3
nnunetv2==2.6.2
notebook_shim==0.2.4
numexpr==2.14.1
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
opencv-python==4.12.0.88
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.1
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydot==4.0.1
pygit2==1.19.1
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.3.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.30.0
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
scmrepo==3.6.1
seaborn==0.13.2
semver==3.0.4
Send2Trash==2.0.0
setuptools==70.2.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
simpleitk==2.5.3
six==1.17.0
smmap==5.0.2
soupsieve==2.8.1
sqltrie==0.11.2
stack-data==0.6.3
surface-distance @ git+https://github.com/deepmind/surface-distance.git@1f805cea446805e559badc80dd27f34175e4173d
sympy==1.14.0
tabulate==0.9.0
tcia_utils==3.2.1
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.12.20
timm==1.0.24
tinycss2==1.4.0
tomlkit==0.13.3
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.4
tqdm==4.67.1
traceback2==1.4.0
traitlets==5.14.3
triton==3.3.1
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
Unidecode==1.4.0
unittest2==1.1.0
uri-template==1.3.0
urllib3==2.6.3
vine==5.1.0
voluptuous==0.16.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
yacs==0.1.8
yarl==1.22.0
zc.lockfile==4.0
